requirements.txt
flask
python-dotenv
nltk
requests
stimson-web-scraper

